Rename: 활동 게시판 생성, 조회, 수정 페이지/activity에서 /acitivity-board로 이동
