+++ frontend/src/pages/activity-board/board/components/create-post-button/index.ts
