+++ frontend/src/pages/activity-board/board/components/hero/Hero.tsx
