+++ frontend/src/pages/activity-board/board/components/hero/index.ts
