+++ frontend/src/pages/activity-board/board/components/index.ts
@@ -1,2 +1,3 @@
 export * from './hero'
 export * from './post-list'
+export * from './create-post-button'
