+++ frontend/src/pages/activity-board/board/components/post-list/index.ts
