+++ frontend/src/pages/activity-board/create-board/components/detail/BoardDetail.tsx
