+++ frontend/src/pages/activity-board/create-board/components/detail/index.ts
