+++ frontend/src/pages/activity-board/create-board/components/hero/Hero.tsx
