+++ frontend/src/pages/activity-board/create-board/components/hero/index.ts
