+++ frontend/src/pages/activity-board/create-board/components/index.ts
