+++ frontend/src/pages/activity-board/edit-board/_components/detail/BoardDetail.tsx
