+++ frontend/src/pages/activity-board/edit-board/_components/hero/index.ts
