+++ frontend/src/pages/activity-board/edit-board/_components/image/EditBoardImage.tsx
