+++ frontend/src/pages/activity/index.ts
@@ -1,5 +1,2 @@
-export * from './board'
-export * from './create-board'
 export * from './default'
-export * from './edit-board'
 export * from './redirect'
